Montgomery encoding changes
