+++ include/common/utils.h
@@ -102,44 +102,4 @@
 		}                                \
 	} while (0)
 
-/**
- * @brief Computes the distance between two torus elements
- *
- * In other words, abs( (a mod 1) - (b mod 1))
- * Mostly used in tests
- *
- * @param a An element in the torus
- * @param b Another element in the torus
- *
- * @returns The (shortest) distance between the two elements in the torus
- */
-double rnx_torus_distance(double a, double b);
-
-/**
- * @brief Computes the distance between two torus elements
- *
- * In other words, abs( (a mod 1) - (b mod 1))
- * Mostly used in tests
- *
- * @param a An element in the torus
- * @param b Another element in the torus
- *
- * @returns The (shortest) distance between the two elements in the torus
- */
-uint64_t tnx_torus_distance(uint64_t a, uint64_t b);
-
-uint64_t u64_round_up_div(uint64_t num, uint64_t step);
-int32_t i32_round_up_div(int32_t num, int32_t step);
-
-/**
- * @brief Generic macro to compute an integer division with rounding up when the remainder is non-zero.
- *
- * @param num The dividend
- * @param step The divisior
- *
- */
-#define INT_ROUND_UP_DIV(num, step) _Generic((num), uint64_t: u64_round_up_div, int32_t: i32_round_up_div)(num, step)
-
-uint64_t next_pow2_log(uint64_t num);
-
 #endif  // UTILS_H
