+++ src/core/glwe/bivariate_polynomial.c
@@ -13,6 +13,7 @@
 #include "backend_rng.h"
 #include "glwe_params.h"
 #include "logger.h"
+#include "math_utils.h"
 #include "maths_structures.h"
 #include "rng.h"
 #include "univariate_polynomial.h"
